Cardano.Crypto.VRF.Praos

Description

Verifiable Random Function (VRF) implemented as FFI wrappers around the implementation in https://github.com/input-output-hk/libsodium

VRFAlgorithm API

data PraosVRF

Key sizes

certSizeVRF :: Int

signKeySizeVRF :: Int

verKeySizeVRF :: Int

vrfKeySizeVRF :: Int

Seed and key generation

data Seed

A random seed, used to derive a key pair.

genSeed :: IO Seed

Generate a random seed. Uses randombytes_buf to create random data.

This function provides an alternative way of generating seeds specifically for the PraosVRF algorithm. Unlike the genKeyPairVRF method, which uses a ByteString-based Seed, this seed generation method bypasses the GHC heap, keeping the seed in C-allocated memory instead.

This provides two advantages: 1. It avoids the overhead of unnecessary GHC-side heap allocations. 2. It avoids leaking the seed via the GHC heap; the Seed type itself takes care of zeroing out its memory upon finalization.

keypairFromSeed :: Seed -> (VerKey, SignKey)

Derive a key pair (Sign + Verify) from a seed.

Conversions

outputBytes :: Output -> ByteString

Convert a proof verification output hash into a ByteString that we can inspect.

proofBytes :: Proof -> ByteString

Convert a proof into a ByteString that we can inspect.

skBytes :: SignKey -> ByteString

Convert a signing key into a ByteString that we can inspect.

vkBytes :: VerKey -> ByteString

Convert a verification key into a ByteString that we can inspect.

skToVerKey :: SignKey -> VerKey

Derive a Verification Key from a Signing Key.

skToSeed :: SignKey -> Seed

Get the seed used to generate a given Signing Key

proofFromBytes :: MonadFail m => ByteString -> m Proof

skFromBytes :: MonadFail m => ByteString -> m SignKey

vkFromBytes :: MonadFail m => ByteString -> m VerKey

Core VRF operations

prove :: SignKey -> ByteString -> Maybe Proof

Construct a proof from a Signing Key and a message. Returns Just the proof on success, Nothing if the signing key could not be decoded.

verify :: VerKey -> Proof -> ByteString -> Maybe Output

Verify a VRF proof and validate the Verification Key. Returns Just a hash of the verification result on success, Nothing if the verification did not succeed.

For a given verification key and message, there are many possible proofs but only one possible output hash.

data family SignKeyVRF v

data family VerKeyVRF v

data family CertVRF v

Internal types

data Proof

A proof, as constructed by the prove function.

data SignKey

Signing key. In this implementation, the signing key is actually a 64-byte value that contains both the 32-byte signing key and the corresponding 32-byte verification key.

data VerKey

Verification key.

data Output

Hashed output of a proof verification, as returned by the verify function.